Design: #60 - 방 생성 버튼 생성 및 위치 조정
1. 방 생성 버튼을 생성하였습니다.
2. 방 생성 모달은 기존의 모달 컴포넌트를 재사용하였습니다.
3. 사이드바의 경우 방의 사이드바와 다르므로 컴포넌트에서 메인 페이지로-
가져왔습니다.
